fixed the issues with redirect loops and finished the connection between the application and the databse. Added a dev function for manually adding passwords to the database for testing. Started setting up routing to add new users
