+++ .gitignore
@@ -71,6 +71,7 @@ yarn-error.log
 /libpeerconnection.log
 testem.log
 /typings
+log
 
 # System files
 .DS_Store
