+++ README.md
@@ -15,3 +15,8 @@
 - Open the project in intellij
 - Build src/main/java/AuthenticationProxyApplication.java
 - Run src/main/java/AuthenticationProxyApplication.java
+
+- ! For the application to run the capstone_user must have permission to SELECT from both tables !
+
+## Angular
+- Every time an update is made to the Angular side of the project, you must rebuild the project with "./mvnw generate-resources" and re-run it
