--- src/main/java/com/authenticationproxy/authenticationproxy/controllers/UiController.java
@@ -1,21 +1,0 @@
-package com.authenticationproxy.authenticationproxy.controllers;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class UiController {
-    @GetMapping("/")
-    public String ui() {
-        return "forward:/login";
-    }
-    @GetMapping("/profile")
-    public String profile() {
-        return "forward:/index.html";
-    }
-
-    @GetMapping("/login")
-    public String login() {
-        return "login";
-    }
-}
