+++ src/main/resources/application.properties
@@ -5,7 +5,7 @@ spring.datasource.username = capstone_user
 spring.datasource.password = ${capstone_user}
 
 spring.jpa.show-sql = true
-spring.jpa.hibernate.ddl-auto = none
+spring.jpa.hibernate.ddl-auto = update
 spring.jpa.generate-ddl = false
 spring.jpa.properties.hibernate.format_sql = true
 spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
